feat: restructure pages and add new components for building search and resources
