+++ website/src/pages/BuildingSearchPage.jsx
@@ -1,0 +1,11 @@
+import { Box, Typography } from "@mui/material";
+
+export function BuildingSearchPage() {
+  return (
+    <Box>
+      <Typography variant="h3" gutterBottom>
+        Make the page!
+      </Typography>
+    </Box>
+  );
+}
