+++ website/src/pages/practice-pages/ResourcesPage.jsx
